Rename stretch.py back to ensemble.py
- File now named after sampling approach (ensemble), not move (stretch)
- Matches class names (EnsembleState, EnsembleInfo)
- Public API remains blackjax.stretch() via alias in mcmc/__init__.py
- Updated test imports to use blackjax.mcmc.ensemble

Rationale: "Ensemble" describes the fundamental multi-walker approach,
while "stretch" is just one possible proposal move. This naming better
reflects the algorithm's core concept and allows for future moves.
